requirements.txt
flask
google-api-python-client
google-auth
google-auth-oauthlib
pandas
psycopg2-binary
